rework competition hooks
